requirements.txt
django==3.2
mysqlclient
Pillow
##metodo de pago
django-paypal